pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "hxse", email = "narutohy@163.com"},
]
dependencies = [
    "fastapi>=0.78.0",
    "pandas>=1.4.3",
    "uvicorn>=0.18.2",
    "Pillow>=9.2.0",
    "requests>=2.28.1",
    "gm>=3.0.142",
    "python-telegram-bot>=13.13",
]
requires-python = ">=3.10,<3.11"
license = {text = "MIT"}

[tool.pdm]

[build-system]
requires = ["pdm-pep517>=0.12.0"]
build-backend = "pdm.pep517.api"
